+++ lifterlms.php
@@ -3,15 +3,15 @@
 * Plugin Name: LifterLMS
 * Plugin URI: https://lifterlms.com/
 * Description: LifterLMS, the #1 WordPress LMS solution, makes it easy to create, sell, and protect engaging online courses.
-* Version: 2.7.11
-* Author: Mark Nelson, Thomas Patrick Levy, codeBOX, LLC
+* Version: 2.7.12
+* Author: codeBOX, Thomas Patrick Levy, Mark Nelson
 * Author URI: http://gocodebox.com
 * Text Domain: lifterlms
 * Domain Path: /languages
 * License:     GPLv2
 * License URI: https://www.gnu.org/licenses/gpl-2.0.html
 * Requires at least: 4.0
-* Tested up to: 4.5
+* Tested up to: 4.6.1
 *
 * @package 		LifterLMS
 * @category 	Core
@@ -35,7 +35,7 @@ require_once 'vendor/autoload.php';
  */
 final class LifterLMS {
 
-	public $version = '2.7.11';
+	public $version = '2.7.12';
 
 	protected static $_instance = null;
 
@@ -87,7 +87,15 @@ final class LifterLMS {
 		add_action( 'init', array( $this, 'integrations' ), 1 );
 		add_action( 'init', array( $this, 'include_template_functions' ) );
 		add_action( 'init', array( 'LLMS_Shortcodes', 'init' ) );
-		add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), array( $this, 'add_action_links' ) );
+		add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), array( $this, 'add_action_links' ), 10, 1 );
+
+		// quick and dirty update note for 3.0 release
+		add_action( 'in_plugin_update_message-lifterlms/lifterlms.php', function( $data ) {
+
+			echo '<br>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<strong>' . sprintf( __( '3.0 is a major update. It is important that you make backups and ensure themes and extensions are 3.0 compatible before upgrading. %sRead more here%s.', 'lifterlms' ), '<a href=" https://lifterlms.com/docs/upgrading-to-lifterlms-3-0/" target="_blank">', '</a>' ) . '</strong>';
+			echo "<script>jQuery( '#lifterlms-update .update-message' ).removeClass( 'notice-warning' ).addClass( 'notice-error' );</script>";
+
+		} );
 
 		// load localization files
 		add_action( 'plugins_loaded', array( $this, 'localize' ) );