+++ CHANGELOG.md
@@ -6,10 +6,20 @@
 + Fixed titles of course closed and open messages on the course restrictions options
 + record a start date for access plans based off when order mobes to complete or active for the first time
 + automatically expire limited access plans
++ gave a quick facelift & unification to a lot of admin panel elements
+
+##### Analytcis
+
++ Charts!
++ Updated styles & interface
 
 ##### bbPress
 
 + Restrict individual forums (and their topics) to LifterLMS Membership levels
+
+##### notices
+
++ Admin notices class
 
 
 = v3.0.0-beta.8 - 2016/09/15 =